requirements.txt
Flask==2.0.1
Flask-Mail==0.9.1
python-dotenv==0.19.0
Werkzeug==2.0.1
gunicorn==20.1.0 